refactor(form): modify styling to add buttons within form for scoping issue
Basically, minor styling changes were done so buttons and the form notice could be inside
header-form id, which is the actual form. Not only does this make the page more uniform as all the
form elements are within the form border itself, but it solves nips the issue of having the tabs be
the entire form, which was going to cause some weird scoping errors with our new multistepform
function. So, everything cool to add in our new logic!